ussd_mental_finlit/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Limiter>=3.3
APScheduler>=3.10
python-dotenv>=1.0
requests>=2.28
# Africa's Talking does not have an official Python SDK; we will use REST API via 'requests'



